Debugging length modifier specs

+++ print_int.c
@@ -10,7 +10,7 @@
 int print_int(va_list integer, flag *f)
 {
 	int length = 0;
-	long int num;
+	int num;
 	unsigned long int n = 0;
 
 	n = va_arg(integer, unsigned long int);
